apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trap-inventory-api"
version = "1.0.0"
description = "TRAP Inventory Management System - Backend API"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "Django>=4.2,<5.0",
    "djangorestframework>=3.14",
    "psycopg2-binary>=2.9",
    "drf-spectacular>=0.27",
    "python-dotenv>=1.0",
    "djangorestframework-simplejwt>=5.3",
    "django-cors-headers>=4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "black>=23.0",
    "isort>=5.12",
    "flake8>=6.0",
]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
